[tox] Fix installation of requirements
Both venv and linters (and its children) environments install
kolla-ansible and thus also install the requirements.
However, they were doing this post-factum and thus without the
constraints pin.
This patch also removes the installation of test-requirements
in venv as it is meant to be used for running the software and
we already have environments for unit tests.
The doc requirements are left in place because docs mention
that ``tox -e venv -- reno`` should work. They should be harmless
but I am open to removing them as well.

Change-Id: I15f1ecc216c9ba81dad740c372d297adf279a945