Rework the Quickstart Guide to make it easier for Newcomers
There are too many ways in one guide for the deployment. This
makes it hard for newcomers to stay at one track and get a working
installation. I reworked this from many possible ways to two.
Only virtual enviroment and non virtual enviroment. For the
developer guide I will make a new patchset.

Signed-off-by: Ramona Beermann <rautenberg@osism.tech>
Change-Id: I0c8739077ad41609596f7687bf2fd3628eb082c5