iscsi: Use assert on checks for readability
assert will also fail when we're not meeting the conditions, makes
clear what we're actually testing, and isn't listed as a skipped task
when the condition is ok.

Change I940b1d3eceb98e16fa366c243672f588b1412d70 dropped CentOS 7,
so now we're just asserting that it's not RedHat.

Change-Id: Iec4e9a6922b67f7c9eb79f580fffbcc8160529e9