Fixes task name in notify module to the actual task name
An OVS-DPDK task had part of its name capitalized,
presumably due to the newly added ansible linter,
but this change was not reflected in the notify
module which calls the task. This results in an
inability to deploy openstack with OVS-DPDK.

This commit fixes the name of the task in the
notify module.

Closes-Bug: #2033413
Change-Id: Iae8a937bb2c35d2eb83e2c4522d829a6cfe25f0d
Signed-off-by: Juan Pablo Suazo <jsuazo@whitestack.com>
(cherry picked from commit 95317e5df28cc6641cb705b4c2dd5c9cf20d5586)