post-2023.1: Remove keystone admin endpoint bits
This backport has been updated to include explicit removal of the admin
endpoint, and should be applied to both Bobcat and Antelope.

In the Zed release kolla-ansible stopped creating the Keystone admin
endpoint by default. Operators were instructed to remove the endpoint
manually or set keystone_create_admin_endpoint to keep it during Zed.

The following condition was used to determine whether to create an admin
endpoint:

  kolla_action == 'upgrade' or keystone_create_admin_endpoint | bool

This was intended to be removed in Antelope, but remained in the code
until removed [1] during the Caracal cycle. This means that if an
operator removes the admin endpoint after upgrading to Zed, Kolla
Ansible will recreate it when upgrading to Antelope or Bobcat.

[1] https://review.opendev.org/c/openstack/kolla-ansible/+/904396

Closes-Bug: #2052298
Change-Id: I27028ffae26a57d510e1a78c38ead2f925396e81
(cherry picked from commit 982c4d5e5f9d0dadefaf10c863fb520b47c46930)