Avoid unnecessary secondary fact gathering failure
Previously, during fact gathering when using the '--limit' argument if a
host inside the limit fails to gather facts, this would cause another
host to fail during delegated fact gathering. This is because the
ansible_facts variable would not be populated for the host.

This change fixes the issue by avoiding delegated fact gathering for
hosts inside the limit.

Partial-Bug: #2041860
Change-Id: I8e321469089f0eb80420b788b51e6d028c7fef4d