Change from PERCONA_SCHEMA to mysql for mariabackup
In 2023.1, mariadb was at version 10.6, which meant that
the PERCONA_SCHEMA.xtrabackup_history table was used to
log historicals for backups.

Starting in 2023.2 onwards, mariadb is at version 10.11, and the
default table used is now mysql.mariadb_backup_history.

Because the mysql database already exists, there is no reason
to try and create it. Instead, we just need to update the defaults
and ensure the permissions get added to the correct database.

Related-Bug: #2061889
Change-Id: If146d8f896c70374884807d42ca0e12df3276d48